Updated to make reserved words compatible with MySQL 8, revamped script loaders, and added new DungeonBalance module along with config settings.
