+++ src/common/Logging/Appender.cpp
@@ -231,6 +231,8 @@ char const* Appender::getLogFilterTypeString(LogFilterType type)
         return "WORLD_QUEST";
     case LOG_FILTER_CHALLENGE:
         return "CHALLENGE";
+    case LOG_FILTER_DUNGEONBALANCE:
+        return "DUNGEONBALANCE";
     default:
         break;
     }
