+++ src/server/game/Entities/Creature/Creature.cpp
@@ -682,7 +682,6 @@ bool Creature::InitEntry(uint32 entry, uint32 /*team*/, const CreatureData* data
     else
         SetObjectScale(cinfo->scale);
 
-
     // checked at loading
     m_defaultMovementType = MovementGeneratorType(cinfo->MovementType);
     if (!m_respawnradius && m_defaultMovementType == RANDOM_MOTION_TYPE)
@@ -2498,13 +2497,13 @@ void Creature::setDeathState(DeathState s)
 
         if (sWorld->getBoolConfig(CONFIG_RESPAWN_FROM_PLAYER_ENABLED))
         {
-            if (_respawnDelay <= 600 && !GetMap()->Instanceable()) // квестовые мобы и прочий шлак
+            if (_respawnDelay <= 600 && !GetMap()->Instanceable()) // quest mobs and other trash
             {
                 uint32 targetCount = GetPlayerFromArea(m_areaId);
                 if (targetCount)
                 {
                     if (targetCount >= sWorld->getIntConfig(CONFIG_RESPAWN_FROM_PLAYER_COUNT))
-                        _respawnDelay /= targetCount; // грубый рассчет, конечно, но лучше уж..
+                        _respawnDelay /= targetCount; // a rough calculation, of course, but better...
 
                     if (_respawnDelay < 10)
                         _respawnDelay = urand(10, 15);
