+++ src/server/game/Server/Packets/MailPackets.h
@@ -226,10 +226,10 @@ namespace WorldPackets
             std::vector<MailNextTimeEntry> Next;
         };
 
-        class NotifyRecievedMail : ServerPacket
+        class NotifyReceivedMail : ServerPacket
         {
         public:
-            NotifyRecievedMail() : ServerPacket(SMSG_NOTIFY_RECEIVED_MAIL, 4) { }
+            NotifyReceivedMail() : ServerPacket(SMSG_NOTIFY_RECEIVED_MAIL, 4) { }
 
             WorldPacket const* Write() override;
