+++ src/server/scripts/EasternKingdoms/easternkingdoms_script_loader.cpp
@@ -1,0 +1,453 @@
+/*
+ * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ // This is where scripts' loading functions should be declared:
+void AddSC_alterac_valley();                 //Alterac Valley
+void AddSC_boss_balinda();
+void AddSC_boss_drekthar();
+void AddSC_boss_galvangar();
+void AddSC_boss_vanndar();
+void AddSC_boss_alizabal();                  //Baradin Hold
+void AddSC_boss_argaloth();
+void AddSC_boss_occuthar();
+void AddSC_instance_baradin_hold();
+
+void AddSC_instance_bastion_of_twilight();//Bastion of Twilight
+void AddSC_boss_theralion_and_valiona();
+void AddSC_boss_sinestra();
+void AddSC_bastion_of_twilight();
+void AddSC_boss_halfus_wyrmbreaker();
+void AddSC_boss_chogall();
+void AddSC_boss_ascendant_council();
+void AddSC_blackrock_depths();               //Blackrock Depths
+void AddSC_boss_ambassador_flamelash();
+void AddSC_boss_anubshiah();
+void AddSC_boss_draganthaurissan();
+void AddSC_boss_general_angerforge();
+void AddSC_boss_gorosh_the_dervish();
+void AddSC_boss_grizzle();
+void AddSC_boss_high_interrogator_gerstahn();
+void AddSC_boss_magmus();
+void AddSC_boss_moira_bronzebeard();
+void AddSC_boss_tomb_of_seven();
+void AddSC_instance_blackrock_depths();
+void AddSC_boss_drakkisath();                //Blackrock Spire
+void AddSC_boss_halycon();
+void AddSC_boss_highlordomokk();
+void AddSC_boss_mothersmolderweb();
+void AddSC_boss_overlordwyrmthalak();
+void AddSC_boss_shadowvosh();
+void AddSC_boss_thebeast();
+void AddSC_boss_warmastervoone();
+void AddSC_boss_quatermasterzigris();
+void AddSC_boss_pyroguard_emberseer();
+void AddSC_boss_gyth();
+void AddSC_boss_rend_blackhand();
+void AddSC_instance_blackrock_spire();
+void AddSC_boss_atramedes();                 //BlackwingDescent
+void AddSC_boss_chimaeron();
+void AddSC_boss_magmaw();
+void AddSC_boss_maloriak();
+void AddSC_boss_bwd_nefarian();
+void AddSC_boss_omnotron_defence_system();
+void AddSC_instance_blackwing_descent();
+void AddSC_blackwing_descent();
+void AddSC_boss_razorgore();                 //Blackwing lair
+void AddSC_boss_vaelastrasz();
+void AddSC_boss_broodlord();
+void AddSC_boss_firemaw();
+void AddSC_boss_ebonroc();
+void AddSC_boss_flamegor();
+void AddSC_boss_chromaggus();
+void AddSC_boss_nefarian();
+void AddSC_instance_blackwing_lair();
+void AddSC_instance_deadmines(); // Deadmines
+void AddSC_deadmines();
+void AddSC_boss_glubtok();
+void AddSC_boss_helix_gearbreaker();
+void AddSC_boss_foereaper5000();
+void AddSC_boss_admiral_ripsnarl();
+void AddSC_boss_captain_cookie();
+void AddSC_boss_vanessa_vancleef();
+void AddSC_gnomeregan();                     //Gnomeregan
+void AddSC_instance_gnomeregan();
+void AddSC_boss_attumen();                   //Karazhan
+void AddSC_boss_curator();
+void AddSC_boss_maiden_of_virtue();
+void AddSC_boss_shade_of_aran();
+void AddSC_boss_malchezaar();
+void AddSC_boss_terestian_illhoof();
+void AddSC_boss_moroes();
+void AddSC_bosses_opera();
+void AddSC_boss_netherspite();
+void AddSC_instance_karazhan();
+void AddSC_karazhan();
+void AddSC_boss_nightbane();
+void AddSC_boss_felblood_kaelthas();         // Magister's Terrace
+void AddSC_boss_selin_fireheart();
+void AddSC_boss_vexallus();
+void AddSC_boss_priestess_delrissa();
+void AddSC_instance_magisters_terrace();
+void AddSC_magisters_terrace();
+void AddSC_boss_lucifron();                  //Molten core
+void AddSC_boss_magmadar();
+void AddSC_boss_gehennas();
+void AddSC_boss_garr();
+void AddSC_boss_baron_geddon();
+void AddSC_boss_shazzrah();
+void AddSC_boss_golemagg();
+void AddSC_boss_sulfuron();
+void AddSC_boss_majordomo();
+void AddSC_boss_ragnaros();
+void AddSC_instance_molten_core();
+void AddSC_the_scarlet_enclave();            //Scarlet Enclave
+void AddSC_the_scarlet_enclave_c1();
+void AddSC_the_scarlet_enclave_c2();
+void AddSC_the_scarlet_enclave_c5();
+void AddSC_boss_thalnos_the_soulrender();    //Scarlet Monastery
+void AddSC_boss_brother_korloff();
+void AddSC_boss_high_inquisitor_whitemane();
+void AddSC_boss_headless_horseman();
+void AddSC_instance_scarlet_monastery();
+void AddSC_boss_armsmaster_harlan();         //Scarlet Halls
+void AddSC_boss_houndmaster_braun();
+void AddSC_boss_flameweaver_koegler();
+void AddSC_instance_scarlet_halls();
+void AddSC_instance_scholomance();           // Update in Mist of Pandaria
+void AddSC_scholomance();
+void AddSC_boss_instructor_chillheart();
+void AddSC_boss_jandice_barov();
+void AddSC_boss_rattlegore();
+void AddSC_boss_lilian_voss();
+void AddSC_boss_darkmaster_gandling();
+void AddSC_boss_baron_ashbury();            //Shadowfang keep
+void AddSC_boss_baron_silverlaine();
+void AddSC_boss_commander_springvale();
+void AddSC_boss_lord_valden();
+void AddSC_boss_lord_godfrey();
+void AddSC_shadowfang_keep();
+void AddSC_instance_shadowfang_keep();
+void AddSC_boss_magistrate_barthilas();      //Stratholme
+void AddSC_boss_maleki_the_pallid();
+void AddSC_boss_nerubenkan();
+void AddSC_boss_cannon_master_willey();
+void AddSC_boss_baroness_anastari();
+void AddSC_boss_ramstein_the_gorger();
+void AddSC_boss_timmy_the_cruel();
+void AddSC_boss_postmaster_malown();
+void AddSC_boss_baron_rivendare();
+void AddSC_boss_dathrohan_balnazzar();
+void AddSC_boss_order_of_silver_hand();
+void AddSC_instance_stratholme();
+void AddSC_stratholme();
+void AddSC_sunken_temple();                  // Sunken Temple
+void AddSC_instance_sunken_temple();
+void AddSC_instance_sunwell_plateau();       //Sunwell Plateau
+void AddSC_boss_kalecgos();
+void AddSC_boss_brutallus();
+void AddSC_boss_felmyst();
+void AddSC_boss_eredar_twins();
+void AddSC_boss_muru();
+void AddSC_boss_kiljaeden();
+void AddSC_sunwell_plateau();
+
+void AddSC_instance_throne_of_the_tides(); // Throne of the Tides
+void AddSC_throne_of_the_tides();
+void AddSC_boss_lady_nazjar();
+void AddSC_boss_commander_ulthok();
+void AddSC_boss_erunak_stonespeaker();
+void AddSC_boss_ozumat();
+
+void AddSC_instance_blackrock_caverns(); // Blackrock Caverns
+void AddSC_boss_romogg_bonecrusher();
+void AddSC_boss_corla_herald_of_twilight();
+void AddSC_boss_karsh_steelbender();
+void AddSC_boss_beauty();
+void AddSC_boss_ascendant_lord_obsidius();
+
+void AddSC_boss_archaedas();                 //Uldaman
+void AddSC_boss_ironaya();
+void AddSC_uldaman();
+void AddSC_instance_uldaman();
+void AddSC_boss_akilzon();                   //Zul'Aman
+void AddSC_boss_halazzi();
+void AddSC_boss_hex_lord_malacrass();
+void AddSC_boss_janalai();
+void AddSC_boss_nalorakk();
+void AddSC_boss_daakara();
+void AddSC_instance_zulaman();
+void AddSC_zulaman();
+void AddSC_instance_grim_batol(); // Grim Batol
+void AddSC_grim_batol();
+void AddSC_boss_general_umbriss();
+void AddSC_boss_forgemaster_throngus();
+void AddSC_boss_drahga_shadowburner();
+void AddSC_boss_erudax();
+void AddSC_boss_grilek();                   // Zul'Gurub
+void AddSC_boss_hazzarah();
+void AddSC_boss_jindo_the_godbreaker();
+void AddSC_boss_kilnara();
+void AddSC_boss_mandokir();
+void AddSC_boss_renataki();
+void AddSC_boss_venoxis();
+void AddSC_boss_wushoolay();
+void AddSC_boss_zanzil();
+void AddSC_instance_zulgurub();
+//void AddSC_alterac_mountains();
+void AddSC_arathi_highlands();
+void AddSC_burning_steppes();
+void AddSC_duskwood();
+void AddSC_eastern_plaguelands();
+void AddSC_eversong_woods();
+void AddSC_ghostlands();
+void AddSC_hinterlands();
+void AddSC_ironforge();
+void AddSC_isle_of_queldanas();
+void AddSC_loch_modan();
+void AddSC_redridge_mountains();
+void AddSC_silvermoon_city();
+void AddSC_silverpine_forest();
+void AddSC_stormwind_city();
+void AddSC_stranglethorn_vale();
+void AddSC_swamp_of_sorrows();
+void AddSC_tirisfal_glades();
+void AddSC_undercity();
+void AddSC_western_plaguelands();
+void AddSC_westfall();
+void AddSC_wetlands();
+
+// The name of this function should match:
+// void Add${NameOfDirectory}Scripts()
+void AddEasternKingdomsScripts()
+{
+    AddSC_alterac_valley();                 //Alterac Valley
+    AddSC_boss_balinda();
+    AddSC_boss_drekthar();
+    AddSC_boss_galvangar();
+    AddSC_boss_vanndar();
+    AddSC_boss_alizabal();                  //Baradin Hold
+    AddSC_boss_argaloth();
+    AddSC_boss_occuthar();
+    AddSC_instance_baradin_hold();
+
+    AddSC_instance_bastion_of_twilight();   //BastionOfTwilight
+    AddSC_boss_theralion_and_valiona();
+    AddSC_boss_sinestra();
+    AddSC_bastion_of_twilight();
+    AddSC_boss_halfus_wyrmbreaker();
+    AddSC_boss_chogall();
+    AddSC_boss_ascendant_council();
+    AddSC_blackrock_depths();               //Blackrock Depths
+    AddSC_boss_ambassador_flamelash();
+    AddSC_boss_anubshiah();
+    AddSC_boss_draganthaurissan();
+    AddSC_boss_general_angerforge();
+    AddSC_boss_gorosh_the_dervish();
+    AddSC_boss_grizzle();
+    AddSC_boss_high_interrogator_gerstahn();
+    AddSC_boss_magmus();
+    AddSC_boss_moira_bronzebeard();
+    AddSC_boss_tomb_of_seven();
+    AddSC_instance_blackrock_depths();
+    AddSC_boss_drakkisath();                //Blackrock Spire
+    AddSC_boss_halycon();
+    AddSC_boss_highlordomokk();
+    AddSC_boss_mothersmolderweb();
+    AddSC_boss_overlordwyrmthalak();
+    AddSC_boss_shadowvosh();
+    AddSC_boss_thebeast();
+    AddSC_boss_warmastervoone();
+    AddSC_boss_quatermasterzigris();
+    AddSC_boss_pyroguard_emberseer();
+    AddSC_boss_gyth();
+    AddSC_boss_rend_blackhand();
+    AddSC_instance_blackrock_spire();
+    AddSC_boss_atramedes();                 //BlackwingDescent
+    AddSC_boss_chimaeron();
+    AddSC_boss_magmaw();
+    AddSC_boss_maloriak();
+    AddSC_boss_bwd_nefarian();
+    AddSC_boss_omnotron_defence_system();
+    AddSC_instance_blackwing_descent();
+    AddSC_blackwing_descent();
+    AddSC_boss_razorgore();                 //Blackwing lair
+    AddSC_boss_vaelastrasz();
+    AddSC_boss_broodlord();
+    AddSC_boss_firemaw();
+    AddSC_boss_ebonroc();
+    AddSC_boss_flamegor();
+    AddSC_boss_chromaggus();
+    AddSC_boss_nefarian();
+    AddSC_instance_blackwing_lair();
+    AddSC_instance_deadmines(); // Deadmines
+    AddSC_deadmines();
+    AddSC_boss_glubtok();
+    AddSC_boss_helix_gearbreaker();
+    AddSC_boss_foereaper5000();
+    AddSC_boss_admiral_ripsnarl();
+    AddSC_boss_captain_cookie();
+    AddSC_boss_vanessa_vancleef();
+    AddSC_gnomeregan();                     //Gnomeregan
+    AddSC_instance_gnomeregan();
+    AddSC_boss_attumen();                   //Karazhan
+    AddSC_boss_curator();
+    AddSC_boss_maiden_of_virtue();
+    AddSC_boss_shade_of_aran();
+    AddSC_boss_malchezaar();
+    AddSC_boss_terestian_illhoof();
+    AddSC_boss_moroes();
+    AddSC_bosses_opera();
+    AddSC_boss_netherspite();
+    AddSC_instance_karazhan();
+    AddSC_karazhan();
+    AddSC_boss_nightbane();
+    AddSC_boss_felblood_kaelthas();         // Magister's Terrace
+    AddSC_boss_selin_fireheart();
+    AddSC_boss_vexallus();
+    AddSC_boss_priestess_delrissa();
+    AddSC_instance_magisters_terrace();
+    AddSC_magisters_terrace();
+    AddSC_boss_lucifron();                  //Molten core
+    AddSC_boss_magmadar();
+    AddSC_boss_gehennas();
+    AddSC_boss_garr();
+    AddSC_boss_baron_geddon();
+    AddSC_boss_shazzrah();
+    AddSC_boss_golemagg();
+    AddSC_boss_sulfuron();
+    AddSC_boss_majordomo();
+    AddSC_boss_ragnaros();
+    AddSC_instance_molten_core();
+    AddSC_the_scarlet_enclave();            //Scarlet Enclave
+    AddSC_the_scarlet_enclave_c1();
+    AddSC_the_scarlet_enclave_c2();
+    AddSC_the_scarlet_enclave_c5();
+    AddSC_boss_thalnos_the_soulrender();    //Scarlet Monastery
+    AddSC_boss_brother_korloff();
+    AddSC_boss_high_inquisitor_whitemane();
+    AddSC_boss_headless_horseman();
+    AddSC_instance_scarlet_monastery();
+    AddSC_boss_armsmaster_harlan();         //Scarlet Halls
+    AddSC_boss_houndmaster_braun();
+    AddSC_boss_flameweaver_koegler();
+    AddSC_instance_scarlet_halls();
+    AddSC_instance_scholomance();           //Scholomance
+    AddSC_boss_instructor_chillheart();
+    AddSC_boss_jandice_barov();
+    AddSC_boss_rattlegore();
+    AddSC_boss_lilian_voss();
+    AddSC_scholomance();
+    AddSC_boss_darkmaster_gandling();
+    AddSC_boss_baron_ashbury();             //Shadowfang keep
+    AddSC_boss_baron_silverlaine();
+    AddSC_boss_commander_springvale();
+    AddSC_boss_lord_valden();
+    AddSC_boss_lord_godfrey();
+    AddSC_shadowfang_keep();
+    AddSC_instance_shadowfang_keep();
+    AddSC_boss_magistrate_barthilas();      //Stratholme
+    AddSC_boss_maleki_the_pallid();
+    AddSC_boss_nerubenkan();
+    AddSC_boss_cannon_master_willey();
+    AddSC_boss_baroness_anastari();
+    AddSC_boss_ramstein_the_gorger();
+    AddSC_boss_timmy_the_cruel();
+    AddSC_boss_postmaster_malown();
+    AddSC_boss_baron_rivendare();
+    AddSC_boss_dathrohan_balnazzar();
+    AddSC_boss_order_of_silver_hand();
+    AddSC_instance_stratholme();
+    AddSC_stratholme();
+    AddSC_sunken_temple();                  // Sunken Temple
+    AddSC_instance_sunken_temple();
+    AddSC_instance_sunwell_plateau();       //Sunwell Plateau
+    AddSC_boss_kalecgos();
+    AddSC_boss_brutallus();
+    AddSC_boss_felmyst();
+    AddSC_boss_eredar_twins();
+    AddSC_boss_muru();
+    AddSC_boss_kiljaeden();
+    AddSC_sunwell_plateau();
+
+    AddSC_instance_throne_of_the_tides(); // Throne of the Tides
+    AddSC_throne_of_the_tides();
+    AddSC_boss_lady_nazjar();
+    AddSC_boss_commander_ulthok();
+    AddSC_boss_erunak_stonespeaker();
+    AddSC_boss_ozumat();
+
+    AddSC_instance_blackrock_caverns(); // Blackrock Caverns
+    AddSC_boss_romogg_bonecrusher();
+    AddSC_boss_corla_herald_of_twilight();
+    AddSC_boss_karsh_steelbender();
+    AddSC_boss_beauty();
+    AddSC_boss_ascendant_lord_obsidius();
+
+    AddSC_boss_archaedas();                 //Uldaman
+    AddSC_boss_ironaya();
+    AddSC_uldaman();
+    AddSC_instance_uldaman();
+    AddSC_boss_akilzon();                   //Zul'Aman
+    AddSC_boss_halazzi();
+    AddSC_boss_hex_lord_malacrass();
+    AddSC_boss_janalai();
+    AddSC_boss_nalorakk();
+    AddSC_boss_daakara();
+    AddSC_instance_zulaman();
+    AddSC_zulaman();
+
+    AddSC_instance_grim_batol(); // Grim Batol
+    AddSC_grim_batol();
+    AddSC_boss_general_umbriss();
+    AddSC_boss_forgemaster_throngus();
+    AddSC_boss_drahga_shadowburner();
+    AddSC_boss_erudax();
+
+    AddSC_boss_grilek();                    // Zul'Gurub
+    AddSC_boss_hazzarah();
+    AddSC_boss_jindo_the_godbreaker();
+    AddSC_boss_kilnara();
+    AddSC_boss_mandokir();
+    AddSC_boss_renataki();
+    AddSC_boss_venoxis();
+    AddSC_boss_wushoolay();
+    AddSC_boss_zanzil();
+    AddSC_instance_zulgurub();
+
+    //AddSC_alterac_mountains();
+    AddSC_arathi_highlands();
+    AddSC_burning_steppes();
+    AddSC_duskwood();
+    AddSC_eastern_plaguelands();
+    AddSC_eversong_woods();
+    AddSC_ghostlands();
+    AddSC_hinterlands();
+    AddSC_ironforge();
+    AddSC_isle_of_queldanas();
+    AddSC_loch_modan();
+    AddSC_redridge_mountains();
+    AddSC_silvermoon_city();
+    AddSC_silverpine_forest();
+    AddSC_stormwind_city();
+    AddSC_stranglethorn_vale();
+    AddSC_swamp_of_sorrows();
+    AddSC_tirisfal_glades();
+    AddSC_undercity();
+    AddSC_western_plaguelands();
+    AddSC_westfall();
+    AddSC_wetlands();
+}
