+++ src/server/scripts/Kalimdor/kalimdor_script_loader.cpp
@@ -1,0 +1,336 @@
+/*
+ * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ // This is where scripts' loading functions should be declared:
+void AddSC_instance_lost_city_of_the_tolvir(); // Lost City of the Tol'Vir
+void AddSC_lost_city_of_the_tolvir();
+void AddSC_boss_general_husam();
+void AddSC_boss_lockmaw_augh();
+void AddSC_boss_high_prophet_barim();
+void AddSC_boss_siamat();
+
+void AddSC_instance_halls_of_origination(); // Halls of Origination
+void AddSC_halls_of_origination();
+void AddSC_boss_temple_guardian_anhuur();
+void AddSC_boss_earthrager_ptah();
+void AddSC_boss_anraphet();
+void AddSC_boss_ammunae();
+void AddSC_boss_isiset();
+void AddSC_boss_setesh();
+void AddSC_boss_rajh();
+
+void AddSC_blackfathom_deeps();              //Blackfathom Depths
+void AddSC_boss_gelihast();
+void AddSC_boss_kelris();
+void AddSC_boss_aku_mai();
+void AddSC_instance_blackfathom_deeps();
+void AddSC_hyjal();                          //CoT Battle for Mt. Hyjal
+void AddSC_boss_archimonde();
+void AddSC_instance_mount_hyjal();
+void AddSC_hyjal_trash();
+void AddSC_boss_rage_winterchill();
+void AddSC_boss_anetheron();
+void AddSC_boss_kazrogal();
+void AddSC_boss_azgalor();
+void AddSC_boss_captain_skarloc();           //CoT Old Hillsbrad
+void AddSC_boss_epoch_hunter();
+void AddSC_boss_lieutenant_drake();
+void AddSC_instance_old_hillsbrad();
+void AddSC_old_hillsbrad();
+void AddSC_boss_aeonus();                    //CoT The Dark Portal
+void AddSC_boss_chrono_lord_deja();
+void AddSC_boss_temporus();
+void AddSC_dark_portal();
+void AddSC_instance_dark_portal();
+
+void AddSC_instance_dragon_soul(); // Dragon Soul
+void AddSC_dragon_soul();
+void AddSC_boss_morchok();
+void AddSC_boss_yorsahj_the_unsleeping();
+void AddSC_boss_warlord_zonozz();
+void AddSC_boss_hagara_the_stormbinder();
+void AddSC_boss_ultraxion();
+void AddSC_boss_warmaster_blackhorn();
+void AddSC_spine_of_deathwing();
+void AddSC_madness_of_deathwing();
+
+void AddSC_instance_end_time(); // End Time
+void AddSC_end_time();
+void AddSC_boss_echo_of_tyrande();
+void AddSC_boss_echo_of_sylvanas();
+void AddSC_boss_echo_of_baine();
+void AddSC_boss_echo_of_jaina();
+void AddSC_boss_murozond();
+
+void AddSC_boss_perotharn();                 //CoT Well of Eternity
+void AddSC_boss_queen_azshara();
+void AddSC_instance_well_of_eternity();
+
+void AddSC_instance_hour_of_twilight(); // Hour of Twilight
+void AddSC_hour_of_twilight();
+void AddSC_boss_arcurion();
+void AddSC_boss_asira_dawnslayer();
+void AddSC_boss_archbishop_benedictus();
+
+void AddSC_boss_epoch();                     //CoT Culling Of Stratholme
+void AddSC_boss_infinite_corruptor();
+void AddSC_boss_salramm();
+void AddSC_boss_mal_ganis();
+void AddSC_boss_meathook();
+void AddSC_culling_of_stratholme();
+void AddSC_instance_culling_of_stratholme();
+void AddSC_boss_celebras_the_cursed();       //Maraudon
+void AddSC_boss_landslide();
+void AddSC_boss_noxxion();
+void AddSC_boss_ptheradras();
+void AddSC_instance_maraudon();
+void AddSC_boss_onyxia();                    //Onyxia's Lair
+void AddSC_instance_onyxias_lair();
+void AddSC_boss_amnennar_the_coldbringer();  //Razorfen Downs
+void AddSC_razorfen_downs();
+void AddSC_instance_razorfen_downs();
+void AddSC_razorfen_kraul();                 //Razorfen Kraul
+void AddSC_instance_razorfen_kraul();
+void AddSC_boss_kurinnaxx();                 //Ruins of ahn'qiraj
+void AddSC_boss_rajaxx();
+void AddSC_boss_moam();
+void AddSC_boss_buru();
+void AddSC_boss_ayamiss();
+void AddSC_boss_ossirian();
+void AddSC_instance_ruins_of_ahnqiraj();
+void AddSC_boss_cthun();                     //Temple of ahn'qiraj
+void AddSC_boss_fankriss();
+void AddSC_boss_huhuran();
+void AddSC_bug_trio();
+void AddSC_boss_sartura();
+void AddSC_boss_skeram();
+void AddSC_boss_twinemperors();
+void AddSC_boss_ouro();
+void AddSC_mob_anubisath_sentinel();
+void AddSC_instance_temple_of_ahnqiraj();
+
+void AddSC_instance_the_vortex_pinnacle(); // The Vortex Pinnacle
+void AddSC_the_vortex_pinnacle();
+void AddSC_boss_grand_vizier_ertan();
+void AddSC_boss_altairus();
+void AddSC_boss_asaad();
+
+void AddSC_throne_of_the_four_winds(); // Throne of Four Winds
+void AddSC_boss_conclave_of_wind();
+void AddSC_boss_alakir();
+void AddSC_instance_throne_of_the_four_winds();
+
+void AddSC_wailing_caverns();                //Wailing caverns
+void AddSC_instance_wailing_caverns();
+void AddSC_zulfarrak();                     //Zul'Farrak generic
+void AddSC_instance_zulfarrak();            //Zul'Farrak instance script
+void AddSC_gilneas();
+
+void AddSC_instance_firelands(); // Firelands
+void AddSC_firelands();
+void AddSC_boss_shannox();
+void AddSC_boss_bethtilac();
+void AddSC_boss_alysrazor();
+void AddSC_boss_lord_rhyolith();
+void AddSC_boss_baleroc();
+void AddSC_boss_majordomo_staghelm();
+void AddSC_boss_ragnaros_firelands();
+
+void AddSC_ashenvale();
+void AddSC_azshara();
+void AddSC_azuremyst_isle();
+void AddSC_bloodmyst_isle();
+void AddSC_darkshore();
+void AddSC_desolace();
+void AddSC_durotar();
+void AddSC_dustwallow_marsh();
+void AddSC_felwood();
+void AddSC_feralas();
+void AddSC_moonglade();
+void AddSC_mulgore();
+void AddSC_orgrimmar();
+void AddSC_silithus();
+void AddSC_stonetalon_mountains();
+void AddSC_tanaris();
+void AddSC_teldrassil();
+void AddSC_the_barrens();
+void AddSC_thousand_needles();
+void AddSC_thunder_bluff();
+void AddSC_ungoro_crater();
+void AddSC_winterspring();
+
+// The name of this function should match:
+// void Add${NameOfDirectory}Scripts()
+void AddKalimdorScripts()
+{
+    AddSC_instance_lost_city_of_the_tolvir(); // Lost City of the Tol'Vir
+    AddSC_lost_city_of_the_tolvir();
+    AddSC_boss_general_husam();
+    AddSC_boss_lockmaw_augh();
+    AddSC_boss_high_prophet_barim();
+    AddSC_boss_siamat();
+
+    AddSC_instance_halls_of_origination(); // Halls of Origination
+    AddSC_halls_of_origination();
+    AddSC_boss_temple_guardian_anhuur();
+    AddSC_boss_earthrager_ptah();
+    AddSC_boss_anraphet();
+    AddSC_boss_ammunae();
+    AddSC_boss_isiset();
+    AddSC_boss_setesh();
+    AddSC_boss_rajh();
+
+    AddSC_blackfathom_deeps();              //Blackfathom Depths
+    AddSC_boss_gelihast();
+    AddSC_boss_kelris();
+    AddSC_boss_aku_mai();
+    AddSC_instance_blackfathom_deeps();
+    AddSC_hyjal();                          //CoT Battle for Mt. Hyjal
+    AddSC_boss_archimonde();
+    AddSC_instance_mount_hyjal();
+    AddSC_hyjal_trash();
+    AddSC_boss_rage_winterchill();
+    AddSC_boss_anetheron();
+    AddSC_boss_kazrogal();
+    AddSC_boss_azgalor();
+    AddSC_boss_captain_skarloc();           //CoT Old Hillsbrad
+    AddSC_boss_epoch_hunter();
+    AddSC_boss_lieutenant_drake();
+    AddSC_instance_old_hillsbrad();
+    AddSC_old_hillsbrad();
+    AddSC_boss_aeonus();                    //CoT The Dark Portal
+    AddSC_boss_chrono_lord_deja();
+    AddSC_boss_temporus();
+    AddSC_dark_portal();
+    AddSC_instance_dark_portal();
+
+    AddSC_instance_dragon_soul(); // Dragon Soul
+    AddSC_dragon_soul();
+    AddSC_boss_morchok();
+    AddSC_boss_yorsahj_the_unsleeping();
+    AddSC_boss_warlord_zonozz();
+    AddSC_boss_hagara_the_stormbinder();
+    AddSC_boss_ultraxion();
+    AddSC_boss_warmaster_blackhorn();
+    AddSC_spine_of_deathwing();
+    AddSC_madness_of_deathwing();
+
+    AddSC_instance_end_time(); // End Time
+    AddSC_end_time();
+    AddSC_boss_echo_of_tyrande();
+    AddSC_boss_echo_of_sylvanas();
+    AddSC_boss_echo_of_baine();
+    AddSC_boss_echo_of_jaina();
+    AddSC_boss_murozond();
+
+    AddSC_boss_perotharn();                 //CoT Well of Eternity
+    AddSC_boss_queen_azshara();
+    AddSC_instance_well_of_eternity();
+
+    AddSC_instance_hour_of_twilight();      // Hour of Twilight
+    AddSC_hour_of_twilight();
+    AddSC_boss_arcurion();
+    AddSC_boss_asira_dawnslayer();
+    AddSC_boss_archbishop_benedictus();
+
+    AddSC_boss_epoch();                     //CoT Culling Of Stratholme
+    AddSC_boss_infinite_corruptor();
+    AddSC_boss_salramm();
+    AddSC_boss_mal_ganis();
+    AddSC_boss_meathook();
+    AddSC_culling_of_stratholme();
+    AddSC_instance_culling_of_stratholme();
+    AddSC_boss_celebras_the_cursed();       //Maraudon
+    AddSC_boss_landslide();
+    AddSC_boss_noxxion();
+    AddSC_boss_ptheradras();
+    AddSC_instance_maraudon();
+    AddSC_boss_onyxia();                    //Onyxia's Lair
+    AddSC_instance_onyxias_lair();
+    AddSC_boss_amnennar_the_coldbringer();  //Razorfen Downs
+    AddSC_razorfen_downs();
+    AddSC_instance_razorfen_downs();
+    AddSC_razorfen_kraul();                 //Razorfen Kraul
+    AddSC_instance_razorfen_kraul();
+    AddSC_boss_kurinnaxx();                 //Ruins of ahn'qiraj
+    AddSC_boss_rajaxx();
+    AddSC_boss_moam();
+    AddSC_boss_buru();
+    AddSC_boss_ayamiss();
+    AddSC_boss_ossirian();
+    AddSC_instance_ruins_of_ahnqiraj();
+    AddSC_boss_cthun();                     //Temple of ahn'qiraj
+    AddSC_boss_fankriss();
+    AddSC_boss_huhuran();
+    AddSC_bug_trio();
+    AddSC_boss_sartura();
+    AddSC_boss_skeram();
+    AddSC_boss_twinemperors();
+    AddSC_boss_ouro();
+    AddSC_mob_anubisath_sentinel();
+    AddSC_instance_temple_of_ahnqiraj();
+
+    AddSC_instance_the_vortex_pinnacle();     // The Vortex Pinnacle
+    AddSC_the_vortex_pinnacle();
+    AddSC_boss_grand_vizier_ertan();
+    AddSC_boss_altairus();
+    AddSC_boss_asaad();
+
+    AddSC_throne_of_the_four_winds();         // Throne of Four Winds
+    AddSC_boss_conclave_of_wind();
+    AddSC_boss_alakir();
+    AddSC_instance_throne_of_the_four_winds();
+
+    AddSC_wailing_caverns();                //Wailing caverns
+    AddSC_instance_wailing_caverns();
+    AddSC_zulfarrak();                      //Zul'Farrak generic
+    AddSC_instance_zulfarrak();             //Zul'Farrak instance script
+    AddSC_gilneas();
+
+    AddSC_instance_firelands();             // Firelands
+    AddSC_firelands();
+    AddSC_boss_shannox();
+    AddSC_boss_bethtilac();
+    AddSC_boss_alysrazor();
+    AddSC_boss_lord_rhyolith();
+    AddSC_boss_baleroc();
+    AddSC_boss_majordomo_staghelm();
+    AddSC_boss_ragnaros_firelands();
+
+    AddSC_ashenvale();
+    AddSC_azshara();
+    AddSC_azuremyst_isle();
+    AddSC_bloodmyst_isle();
+    AddSC_darkshore();
+    AddSC_desolace();
+    AddSC_durotar();
+    AddSC_dustwallow_marsh();
+    AddSC_felwood();
+    AddSC_feralas();
+    AddSC_moonglade();
+    AddSC_mulgore();
+    AddSC_orgrimmar();
+    AddSC_silithus();
+    AddSC_stonetalon_mountains();
+    AddSC_tanaris();
+    AddSC_teldrassil();
+    AddSC_the_barrens();
+    AddSC_thousand_needles();
+    AddSC_thunder_bluff();
+    AddSC_ungoro_crater();
+    AddSC_winterspring();
+}
