+++ src/server/scripts/Legion/legion_script_loader.cpp
@@ -1,0 +1,367 @@
+/*
+ * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ // This is where scripts' loading functions should be declared:
+void AddSC_Mardum();
+void AddSC_warden_prison();
+void AddSC_world_bossess_legion();
+void AddSC_invasion_point_world_bosses();
+void AddSC_invasion_point_argus();
+void AddSC_sentinax();
+void Addsc_paraxis();
+
+void AddSC_brokenIslands();
+void AddSC_instance_broken_islands();
+
+void AddSC_instance_black_rook_hold_dungeon();  //Black Rook Hold Dungeon
+void AddSC_boss_the_amalgam_of_souls();
+void AddSC_boss_illysanna_ravencrest();
+void AddSC_boss_smashspite_the_hateful();
+void AddSC_boss_lord_kurtalos_ravencrest();
+
+void AddSC_instance_darkheart_thicket();       //Dark Heart Thicket
+void AddSC_boss_arch_druid_glaidalis();
+void AddSC_boss_oakheart();
+void AddSC_boss_dresaron();
+void AddSC_boss_shade_of_xavius();
+
+void AddSC_instance_eye_of_azshara();           //Eye of Azshara
+void AddSC_boss_warlord_parjesh();
+void AddSC_boss_lady_hatecoil();
+void AddSC_boss_king_deepbeard();
+void AddSC_boss_serpentrix();
+void AddSC_boss_wrath_of_azshara();
+
+void AddSC_instance_halls_of_valor();           //Halls of Valor
+void AddSC_boss_odyn();
+void AddSC_boss_hyrja();
+void AddSC_boss_hymdall();
+void AddSC_boss_fenryr();
+void AddSC_boss_god_king_skovald();
+void AddSC_instance_maw_of_souls();
+void AddSC_boss_ymiron_the_fallen_king();
+void AddSC_boss_harbaron();
+void AddSC_boss_helya();
+
+void AddSC_instance_neltharions_lair();         //Neltharions Lair
+void AddSC_boss_naraxas();
+void AddSC_boss_rokmora();
+void AddSC_boss_ularogg_cragshaper();
+void AddSC_boss_dargrul_the_underking();
+
+void AddSC_instance_the_arcway();               //The Arcway
+void AddSC_boss_ivanyr();
+void AddSC_boss_corstilax();
+void AddSC_boss_general_xakal();
+void AddSC_boss_naltira();
+void AddSC_boss_advisor_vandros();
+
+void AddSC_instance_vault_of_the_wardens();     //Vault of the Wardens
+void AddSC_boss_tirathon_saltheril();
+void AddSC_boss_inquisitor_tormentorum();
+void AddSC_boss_ashgolm();
+void AddSC_boss_glazer();
+void AddSC_boss_cordana();
+
+void AddSC_instance_violet_hold_legion();       //Violet Hold Legion
+void AddSC_boss_mindflayer_kaahrj();
+void AddSC_boss_millificent_manastorm();
+void AddSC_boss_festerface();
+void AddSC_boss_shivermaw();
+void AddSC_boss_anubesset();
+void AddSC_boss_saelorn();
+void AddSC_boss_blood_princess_thalena();
+void AddSC_boss_fel_lord_betrug();
+void AddSC_violet_hold_legion();
+
+void AddSC_instance_court_of_stars();           //Court of Stars
+void AddSC_boss_patrol_captain_gerdo();
+void AddSC_boss_talixae_flamewreath();
+void AddSC_boss_advisor_melandrus();
+
+void AddSC_instance_return_to_karazhan();       //Return to Karazhan
+void AddSC_boss_opera_hall();
+void AddSC_boss_maiden_of_virtue_legion();
+void AddSC_boss_attumen_the_huntsman();
+void AddSC_boss_moroes_legion();
+void AddSC_boss_the_curator();
+void AddSC_boss_mana_devourer();
+void AddSC_boss_shade_of_medivh();
+void AddSC_boss_vizaduum_the_watcher();
+void AddSC_karazhan_chess();
+void AddSC_boss_rtk_nightbane();
+
+void AddSC_instance_cathedral_of_eternal_night(); //Cathedral of Eternal Night
+void AddSC_boss_agronox();
+void AddSC_boss_thrashbite_the_scornful();
+void AddSC_boss_domatrax();
+void AddSC_boss_mephistroth();
+void AddSC_cathedral_of_eternal_night();
+
+void AddSC_instance_the_emerald_nightmare();      //The Emerald Nightmare
+void AddSC_the_emerald_nightmare();
+void AddSC_boss_nythendra();
+void AddSC_boss_elerethe_renferal();
+void AddSC_boss_ilgynoth();
+void AddSC_boss_ursoc();
+void AddSC_boss_dragons_of_nightmare();
+void AddSC_boss_cenarius();
+void AddSC_boss_xavius();
+
+void AddSC_trial_of_valor();                      //Trial of Valor
+void AddSC_instance_trial_of_valor();
+void AddSC_tov_boss_odyn();
+void AddSC_boss_garm();
+void AddSC_boss_tov_helya();
+
+void AddSC_boss_chronomatic_anomaly();
+void AddSC_boss_elisande();
+void AddSC_boss_guldan();
+void AddSC_boss_high_botanist_telarn();
+void AddSC_boss_krosus();
+void AddSC_boss_skorpyron();
+void AddSC_boss_spellblade_aluriel();
+void AddSC_boss_star_augur_etraeus();
+void AddSC_boss_tichondrius();
+void AddSC_boss_trilliax();
+void AddSC_instance_the_nightnold();
+void AddSC_the_nighthold();
+
+void AddSC_instance_tomb_of_sargeras();
+void AddSC_boss_goroth();
+void AddSC_boss_mistress_sasszine();
+void AddSC_boss_sisters_of_the_moon();
+void AddSC_tomb_of_sargeras();
+void AddSC_boss_demonic_inquisition();
+void AddSC_boss_harjatan();
+void AddSC_boss_maiden_of_vigilance();
+void AddSC_boss_fallen_avatar();
+void AddSC_boss_the_desolate_host();
+void AddSC_boss_tos_kiljaeden();
+
+void AddSC_instance_seat_of_the_triumvirate(); // Seat Of The Triumvirate
+void AddSC_boss_zuraal_the_ascended();
+void AddSC_boss_saprish();
+void AddSC_boss_viceroy_nezhar();
+void AddSC_boss_lura();
+
+void AddSC_instance_antorus();
+void AddSC_boss_worldbreaker();
+void AddSC_boss_felhounds();
+void AddSC_boss_antoran();
+void AddSC_boss_hasabel();
+void AddSC_boss_eonar();
+void AddSC_boss_imonar();
+void AddSC_boss_kingaroth();
+void AddSC_boss_varimathras();
+void AddSC_boss_coven_shivarres();
+void AddSC_boss_aggramar();
+void AddSC_boss_argus();
+void AddSC_antorus();
+
+// Micro-Holidays
+void AddSC_CallOfTheScarab();
+void AddSC_GlowcapFestival();
+void AddSC_HatchingOfTheHippogryphs();
+void AddSC_KirinTorTavernCrawl();
+void AddSC_MarchOfTheTadpoles();
+void AddSC_SpringBalloonFestival();
+void AddSC_ThousandBoatBash();
+void AddSC_UnGoroMadness();
+void AddSC_VolunteerGuardDay();
+void AddSC_TrialofStyle();
+void AddSC_AuctionHouseDanceParty();
+void AddSC_DarkMoonConcertBlightBoar();
+void AddSC_TheGreatGromereganRun();
+void AddSC_MoonkinFestival();
+
+// The name of this function should match:
+// void Add${NameOfDirectory}Scripts()
+void AddLegionScripts()
+{
+    AddSC_Mardum();
+    AddSC_warden_prison();
+    AddSC_world_bossess_legion();
+    AddSC_invasion_point_world_bosses();
+    AddSC_invasion_point_argus();
+    AddSC_sentinax();
+    Addsc_paraxis();
+
+    AddSC_brokenIslands();
+    AddSC_instance_broken_islands();
+
+    AddSC_instance_black_rook_hold_dungeon();  //Black Rook Hold Dungeon
+    AddSC_boss_the_amalgam_of_souls();
+    AddSC_boss_illysanna_ravencrest();
+    AddSC_boss_smashspite_the_hateful();
+    AddSC_boss_lord_kurtalos_ravencrest();
+
+    AddSC_instance_darkheart_thicket();       //Dark Heart Thicket
+    AddSC_boss_arch_druid_glaidalis();
+    AddSC_boss_oakheart();
+    AddSC_boss_dresaron();
+    AddSC_boss_shade_of_xavius();
+
+    AddSC_instance_eye_of_azshara();           //Eye of Azshara
+    AddSC_boss_warlord_parjesh();
+    AddSC_boss_lady_hatecoil();
+    AddSC_boss_king_deepbeard();
+    AddSC_boss_serpentrix();
+    AddSC_boss_wrath_of_azshara();
+
+    AddSC_instance_halls_of_valor();           //Halls of Valor
+    AddSC_boss_odyn();
+    AddSC_boss_hyrja();
+    AddSC_boss_hymdall();
+    AddSC_boss_fenryr();
+    AddSC_boss_god_king_skovald();
+    AddSC_instance_maw_of_souls();
+    AddSC_boss_ymiron_the_fallen_king();
+    AddSC_boss_harbaron();
+    AddSC_boss_helya();
+
+    AddSC_instance_tomb_of_sargeras();
+    AddSC_boss_goroth();
+    AddSC_boss_mistress_sasszine();
+    AddSC_boss_sisters_of_the_moon();
+    AddSC_tomb_of_sargeras();
+    AddSC_boss_demonic_inquisition();
+    AddSC_boss_harjatan();
+    AddSC_boss_maiden_of_vigilance();
+    AddSC_boss_fallen_avatar();
+    AddSC_boss_the_desolate_host();
+    AddSC_boss_tos_kiljaeden();
+
+    AddSC_instance_neltharions_lair();         //Neltharions Lair
+    AddSC_boss_naraxas();
+    AddSC_boss_rokmora();
+    AddSC_boss_ularogg_cragshaper();
+    AddSC_boss_dargrul_the_underking();
+
+    AddSC_instance_the_arcway();               //The Arcway
+    AddSC_boss_ivanyr();
+    AddSC_boss_corstilax();
+    AddSC_boss_general_xakal();
+    AddSC_boss_naltira();
+    AddSC_boss_advisor_vandros();
+
+    AddSC_instance_vault_of_the_wardens();     //Vault of the Wardens
+    AddSC_boss_tirathon_saltheril();
+    AddSC_boss_inquisitor_tormentorum();
+    AddSC_boss_ashgolm();
+    AddSC_boss_glazer();
+    AddSC_boss_cordana();
+
+    AddSC_instance_violet_hold_legion();       //Violet Hold Legion
+    AddSC_boss_mindflayer_kaahrj();
+    AddSC_boss_millificent_manastorm();
+    AddSC_boss_festerface();
+    AddSC_boss_shivermaw();
+    AddSC_boss_anubesset();
+    AddSC_boss_saelorn();
+    AddSC_boss_blood_princess_thalena();
+    AddSC_boss_fel_lord_betrug();
+    AddSC_violet_hold_legion();
+
+    AddSC_instance_court_of_stars();           //Court of Stars
+    AddSC_boss_patrol_captain_gerdo();
+    AddSC_boss_talixae_flamewreath();
+    AddSC_boss_advisor_melandrus();
+
+    AddSC_instance_return_to_karazhan();       //Return to Karazhan
+    AddSC_boss_opera_hall();
+    AddSC_boss_maiden_of_virtue_legion();
+    AddSC_boss_attumen_the_huntsman();
+    AddSC_boss_moroes_legion();
+    AddSC_boss_the_curator();
+    AddSC_boss_mana_devourer();
+    AddSC_boss_shade_of_medivh();
+    AddSC_boss_vizaduum_the_watcher();
+    AddSC_karazhan_chess();
+    AddSC_boss_rtk_nightbane();
+
+    AddSC_instance_cathedral_of_eternal_night(); //Cathedral of Eternal Night
+    AddSC_boss_agronox();
+    AddSC_boss_thrashbite_the_scornful();
+    AddSC_boss_domatrax();
+    AddSC_boss_mephistroth();
+    AddSC_cathedral_of_eternal_night();
+
+    AddSC_instance_the_emerald_nightmare();     //The Emerald Nightmare
+    AddSC_the_emerald_nightmare();
+    AddSC_boss_nythendra();
+    AddSC_boss_elerethe_renferal();
+    AddSC_boss_ilgynoth();
+    AddSC_boss_ursoc();
+    AddSC_boss_dragons_of_nightmare();
+    AddSC_boss_cenarius();
+    AddSC_boss_xavius();
+
+    AddSC_trial_of_valor();                     //Trial of Valor
+    AddSC_instance_trial_of_valor();
+    AddSC_tov_boss_odyn();
+    AddSC_boss_garm();
+    AddSC_boss_tov_helya();
+
+    AddSC_boss_chronomatic_anomaly();           //The Nighthold
+    AddSC_boss_elisande();
+    AddSC_boss_guldan();
+    AddSC_boss_high_botanist_telarn();
+    AddSC_boss_krosus();
+    AddSC_boss_skorpyron();
+    AddSC_boss_spellblade_aluriel();
+    AddSC_boss_star_augur_etraeus();
+    AddSC_boss_tichondrius();
+    AddSC_boss_trilliax();
+    AddSC_instance_the_nightnold();
+    AddSC_the_nighthold();
+
+    AddSC_CallOfTheScarab();
+    AddSC_GlowcapFestival();
+    AddSC_HatchingOfTheHippogryphs();
+    AddSC_KirinTorTavernCrawl();
+    AddSC_MarchOfTheTadpoles();
+    AddSC_SpringBalloonFestival();
+    AddSC_ThousandBoatBash();
+    AddSC_UnGoroMadness();
+    AddSC_VolunteerGuardDay();
+    AddSC_TrialofStyle();
+    AddSC_AuctionHouseDanceParty();
+    AddSC_DarkMoonConcertBlightBoar();
+    AddSC_TheGreatGromereganRun();
+    AddSC_MoonkinFestival();
+
+    AddSC_instance_seat_of_the_triumvirate(); // Seat Of The Triumvirate
+    AddSC_boss_zuraal_the_ascended();
+    AddSC_boss_saprish();
+    AddSC_boss_viceroy_nezhar();
+    AddSC_boss_lura();
+
+    AddSC_instance_antorus();
+    AddSC_boss_worldbreaker();
+    AddSC_boss_felhounds();
+    AddSC_boss_antoran();
+    AddSC_boss_hasabel();
+    AddSC_boss_eonar();
+    AddSC_boss_imonar();
+    AddSC_boss_kingaroth();
+    AddSC_boss_varimathras();
+    AddSC_boss_coven_shivarres();
+    AddSC_boss_aggramar();
+    AddSC_boss_argus();
+    AddSC_antorus();
+}
