+++ src/server/scripts/Pandaria/pandaria_script_loader.cpp
@@ -1,0 +1,221 @@
+/*
+ * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ // This is where scripts' loading functions should be declared:
+void AddSC_instance_temple_of_jade_serpent();       // Temple of Jade Serpent
+void AddSC_boss_wise_mari();
+void AddSC_boss_lorewalker_stonestep();
+void AddSC_boss_liu_flameheat();
+void AddSC_boss_sha_of_doubt();
+void AddSC_instance_stormstout_brewery();           // The Stormstout Brewery
+void AddSC_stormstout_brewery();
+void AddSC_boss_ook_ook();
+void AddSC_boss_hoptallus();
+void AddSC_boss_yan_zhu();
+void AddSC_instance_gate_setting_sun();             // Gate of the Setting Sun
+void AddSC_gate_setting_sun();
+void AddSC_boss_saboteur_kiptilak();
+void AddSC_boss_striker_gadok();
+void AddSC_boss_commander_rimok();
+void AddSC_boss_raigonn();
+void AddSC_boss_sha_of_anger();                     // Pandaria World Bosses
+void AddSC_boss_galion();
+void AddSC_boss_oondasta();
+void AddSC_boss_nalak();
+void AddSC_instance_mogu_shan_palace();             // Mogu'Shan Palace
+void AddSC_mogu_shan_palace();
+void AddSC_boss_trial_of_the_king();
+void AddSC_boss_gekkan();
+void AddSC_boss_xin_the_weaponmaster();
+void AddSC_instance_shadopan_monastery();           // Shadopan Monastery
+void AddSC_shadopan_monastery();
+void AddSC_boss_gu_cloudstrike();
+void AddSC_boss_master_snowdrift();
+void AddSC_boss_sha_of_violence();
+void AddSC_boss_taran_zhu();
+void AddSC_instance_siege_of_the_niuzoa_temple();   // Siege of the Niuzoa Temple
+void AddSC_siege_of_the_niuzoa_temple();
+void AddSC_boss_jinbak();
+void AddSC_boss_commander_vojak();
+void AddSC_boss_general_pavalak();
+void AddSC_boss_wing_leader_neronok();
+void AddSC_instance_mogu_shan_vault();              // Mogu'Shan Vault
+void AddSC_mogu_shan_vault();
+void AddSC_boss_stone_guard();
+void AddSC_boss_feng();
+void AddSC_boss_garajal();
+void AddSC_boss_spirit_kings();
+void AddSC_boss_elegon();
+void AddSC_boss_will_of_emperor();
+
+void AddSC_instance_heart_of_fear();                // Heart of Fear
+void AddSC_boss_vizier_zorlok();
+void AddSC_boss_lord_tayak();
+void AddSC_boss_garalon();
+void AddSC_boss_lord_meljarak();
+void AddSC_boss_unsok();
+void AddSC_boss_empress_shekzeer();
+
+// MoP raids - Terrace of Endless Spring
+void AddSC_instance_terrace_of_endless_spring();
+void AddSC_terrace_of_endless_spring();
+void AddSC_boss_protectors_of_the_endless();
+void AddSC_boss_tsulong();
+void AddSC_boss_lei_shi();
+void AddSC_boss_sha_of_fear();
+void AddSC_instance_throne_of_thunder();            // Throne of Thunder
+void AddSC_boss_jinrokh();
+void AddSC_boss_horridon();
+void AddSC_boss_council_of_elders();
+void AddSC_boss_tortos();
+void AddSC_boss_megaera();
+void AddSC_boss_jikun();
+void AddSC_boss_durumu();
+void AddSC_boss_primordius();
+void AddSC_boss_dark_animus();
+void AddSC_boss_iron_qon();
+void AddSC_boss_twin_consorts();
+void AddSC_boss_lei_shen();
+void AddSC_boss_ra_den();
+void AddSC_instance_siege_of_orgrimmar();           // Siege of Orgrimmar 
+void AddSC_siege_of_orgrimmar();
+void AddSC_boss_immerseus();
+void AddSC_boss_fallen_protectors();
+void AddSC_boss_norushen();
+void AddSC_boss_sha_of_pride();
+void AddSC_boss_galakras();
+void AddSC_boss_iron_juggernaut();
+void AddSC_boss_korkron_dark_shaman();
+void AddSC_boss_general_nazgrim();
+void AddSC_boss_malkorok();
+void AddSC_boss_spoils_of_pandaria();
+void AddSC_boss_thok_the_bloodthirsty();
+void AddSC_boss_siegecrafter_blackfuse();
+void AddSC_boss_paragons_of_the_klaxxi();
+void AddSC_boss_garrosh_hellscream();
+
+void AddSC_valley_of_the_four_winds();
+void AddSC_krasarang_wilds();
+void AddSC_kun_lai_summit();
+
+void AddSC_WanderingIsland();
+void AddSC_timeless_isle();
+
+// The name of this function should match:
+// void Add${NameOfDirectory}Scripts()
+void AddPandariaScripts()
+{
+    AddSC_instance_temple_of_jade_serpent();        // Temple of Jade Serpent
+    AddSC_boss_wise_mari();
+    AddSC_boss_lorewalker_stonestep();
+    AddSC_boss_liu_flameheat();
+    AddSC_boss_sha_of_doubt();
+    AddSC_instance_stormstout_brewery();            // The Stormstout Brewery
+    AddSC_stormstout_brewery();
+    AddSC_boss_ook_ook();
+    AddSC_boss_hoptallus();
+    AddSC_boss_yan_zhu();
+    AddSC_instance_mogu_shan_palace();              // The Mogu'Shan Palace
+    AddSC_mogu_shan_palace();
+    AddSC_boss_trial_of_the_king();
+    AddSC_boss_gekkan();
+    AddSC_boss_xin_the_weaponmaster();
+    AddSC_instance_gate_setting_sun();              // Gate of the Setting Sun
+    AddSC_gate_setting_sun();
+    AddSC_boss_saboteur_kiptilak();
+    AddSC_boss_striker_gadok();
+    AddSC_boss_commander_rimok();
+    AddSC_boss_raigonn();
+    AddSC_boss_sha_of_anger();                      // Pandaria World Bosses
+    AddSC_boss_galion();
+    AddSC_boss_oondasta();
+    AddSC_boss_nalak();
+    AddSC_instance_shadopan_monastery();            // Shadopan Monastery
+    AddSC_shadopan_monastery();
+    AddSC_boss_gu_cloudstrike();
+    AddSC_boss_master_snowdrift();
+    AddSC_boss_sha_of_violence();
+    AddSC_boss_taran_zhu();
+    AddSC_instance_siege_of_the_niuzoa_temple();    // Siege of the Niuzoa Temple
+    AddSC_siege_of_the_niuzoa_temple();
+    AddSC_boss_jinbak();
+    AddSC_boss_commander_vojak();
+    AddSC_boss_general_pavalak();
+    AddSC_boss_wing_leader_neronok();
+    AddSC_instance_mogu_shan_vault();               // Mogu'Shan Vault
+    AddSC_mogu_shan_vault();
+    AddSC_boss_stone_guard();
+    AddSC_boss_feng();
+    AddSC_boss_garajal();
+    AddSC_boss_spirit_kings();
+    AddSC_boss_elegon();
+    AddSC_boss_will_of_emperor();
+
+    AddSC_instance_heart_of_fear();                //Heart of Fear
+    AddSC_boss_vizier_zorlok();
+    AddSC_boss_lord_tayak();
+    AddSC_boss_garalon();
+    AddSC_boss_lord_meljarak();
+    AddSC_boss_unsok();
+    AddSC_boss_empress_shekzeer();
+
+    // Pandaria Raids - Terrace of Endless Spring
+    AddSC_instance_terrace_of_endless_spring();
+    AddSC_terrace_of_endless_spring();
+    AddSC_boss_protectors_of_the_endless();
+    AddSC_boss_tsulong();
+    AddSC_boss_lei_shi();
+    AddSC_boss_sha_of_fear();
+
+    AddSC_instance_throne_of_thunder();            // Throne of Thunder
+    AddSC_boss_jinrokh();
+    AddSC_boss_horridon();
+    AddSC_boss_council_of_elders();
+    AddSC_boss_tortos();
+    AddSC_boss_megaera();
+    AddSC_boss_jikun();
+    AddSC_boss_durumu();
+    AddSC_boss_primordius();
+    AddSC_boss_dark_animus();
+    AddSC_boss_iron_qon();
+    AddSC_boss_twin_consorts();
+    AddSC_boss_lei_shen();
+    AddSC_boss_ra_den();
+    AddSC_instance_siege_of_orgrimmar();           // Siege of Orgrimmar 
+    AddSC_siege_of_orgrimmar();
+    AddSC_boss_immerseus();
+    AddSC_boss_fallen_protectors();
+    AddSC_boss_norushen();
+    AddSC_boss_sha_of_pride();
+    AddSC_boss_galakras();
+    AddSC_boss_iron_juggernaut();
+    AddSC_boss_korkron_dark_shaman();
+    AddSC_boss_general_nazgrim();
+    AddSC_boss_malkorok();
+    AddSC_boss_spoils_of_pandaria();
+    AddSC_boss_thok_the_bloodthirsty();
+    AddSC_boss_siegecrafter_blackfuse();
+    AddSC_boss_paragons_of_the_klaxxi();
+    AddSC_boss_garrosh_hellscream();
+
+    AddSC_valley_of_the_four_winds();
+    AddSC_krasarang_wilds();
+    AddSC_kun_lai_summit();
+
+    AddSC_WanderingIsland();
+    AddSC_timeless_isle();
+}
