+++ src/server/scripts/Scenario/scenario_script_loader.cpp
@@ -1,0 +1,359 @@
+/*
+ * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ // This is where scripts' loading functions should be declared:
+
+//< Arena of Annihilation
+void AddSC_instance_arena_of_annihilation();
+void AddSC_boss_scar_shell();
+void AddSC_boss_jolgrum();
+void AddSC_boss_little_liuyang();
+void AddSC_boss_chagan_firehoof();
+void AddSC_arena_of_annihilation();
+
+//< Fall of Shan'bu
+void AddSC_fall_of_shan_bu();
+void AddSC_instance_fall_of_shan_bu();
+
+//< Pursuing the Black Harvest
+void AddSC_pursing_the_black_harvest();
+void AddSC_instance_pursuing_the_black_harvest();
+
+//< Thunder Forge
+void AddSC_thunder_forge();
+void AddSC_instance_thunder_forge();
+
+//< Troves of the Thunder King
+void AddSC_troves_of_the_thunder_king();
+void AddSC_instance_troves_of_the_thunder_king();
+
+//< A Brewing Storm
+void AddSC_a_brewing_storm();
+void AddSC_instance_a_brewing_storm();
+
+//< A Little Patience
+void AddSC_a_little_patience();
+void AddSC_instance_a_little_patience();
+
+//< Assault on Zan'vess
+void AddSC_assault_on_zanvess();
+void AddSC_instance_assault_on_zanvess();
+
+//< Battle on the High Seas
+void AddSC_battle_on_the_high_seas();
+void AddSC_instance_battle_on_the_high_seas();
+
+//< Blood in the Snow
+void AddSC_blood_in_the_snow();
+void AddSC_instance_blood_in_the_snow();
+
+//< Brewmoon Festival
+void AddSC_brewmoon_festival();
+void AddSC_instance_brewmoon_festival();
+
+//< Celestial Tournament
+void AddSC_celestial_tournament();
+void AddSC_instance_celestial_tournament();
+
+//< Crypt of Forgotten Kings
+void AddSC_crypt_of_forgotten_kings();
+void AddSC_instance_crypt_of_forgotten_kings();
+
+//< Dagger in the Dark
+void AddSC_dagger_in_the_dark();
+void AddSC_instance_dagger_in_the_dark();
+
+//< Dark Heart of Pandaria
+void AddSC_dark_heart_of_pandaria();
+void AddSC_instance_dark_heart_of_pandaria();
+
+//< Domination Point
+void AddSC_domination_point();
+void AddSC_instance_domination_point();
+
+//< Greenstone Village
+void AddSC_greenstone_village();
+void AddSC_instance_greenstone_village();
+
+//< Lion's Landing
+void AddSC_lions_landing();
+void AddSC_instance_lions_landing();
+
+//< The Secrets of Ragefire
+void AddSC_the_secrets_of_ragefire();
+void AddSC_instance_the_secrets_of_ragefire();
+
+//< Unga Ingoo
+void AddSC_unga_ingoo();
+void AddSC_instance_unga_ingoo();
+
+//< Proving Grounds
+void AddSC_instance_proving_grounds();
+void AddSC_proving_grounds();
+
+//< Broken Shore
+void AddSC_instance_broken_shore();
+void AddSC_broken_shore();
+
+//< Shield's Reset
+void AddSC_instance_shields_rest();
+void AddSC_shields_rest();
+
+//< Vengeance Art
+void AddSC_DH_vengeance_art_scenario();
+
+// Rogue Arts
+void AddSC_Dreadblades();
+
+void AddSC_Kingslayers();
+void AddSC_instance_kingslayers();
+
+void AddSC_Devourer();
+void AddSC_instance_devourer();
+
+// Warrior scenarios
+void AddSC_wars_intro_scenario();
+
+void AddSC_Warswords();
+void AddSC_instance_warswords();
+
+void AddSC_EarthWarder();
+
+//priest art
+void AddSC_Tuure();
+void AddSC_instance_tuure();
+
+// Paladin Art
+void AddSC_SilverHand();
+void AddSC_instance_silver_hand();
+
+// Shaman Art
+void AddSC_Sharasdal();
+void AddSC_instance_sharasdal();
+
+// Monk Art
+void AddSC_monk_intro_scenario();
+
+void AddSC_Sheylun();
+
+// Arts common
+void AddSC_instance_class_art_scenarios();
+
+void AddSC_AssaultBrokenShore();
+void AddSC_instance_AssaultBrokenShore();
+
+void AddSC_InvasionAzsuna();
+void AddSC_instance_InvasionAzsuna();
+
+void AddSC_InvasionHM();
+void AddSC_instance_invasion_HM();
+
+void AddSC_InvasionValshara();
+void AddSC_instance_invasion_valshara();
+
+void AddSC_InvasionStormheim();
+void AddSC_instance_invasion_stormheim();
+
+void AddSC_instance_walling_caverns_pet();
+void AddSC_instance_dead_mines_per();
+
+void AddSC_army_training();
+void AddSC_instance_army_training();
+
+void AddSC_instance_temple_of_the_jade_serpent();
+
+void AddSC_brawlers_guild();
+void AddSC_brawlers_guild_bosses_rank_one();
+void AddSC_brawlers_guild_bosses_rank_two();
+void AddSC_brawlers_guild_bosses_rank_three();
+void AddSC_brawlers_guild_bosses_rank_four();
+void AddSC_brawlers_guild_bosses_rank_five();
+void AddSC_brawlers_guild_bosses_rank_six();
+void AddSC_brawlers_guild_bosses_rank_seven();
+
+// The name of this function should match:
+// void Add${NameOfDirectory}Scripts()
+void AddScenarioScripts()
+{
+    //< Arena of Annihilation
+    AddSC_instance_arena_of_annihilation();
+    AddSC_boss_scar_shell();
+    AddSC_boss_jolgrum();
+    AddSC_boss_little_liuyang();
+    AddSC_boss_chagan_firehoof();
+    AddSC_arena_of_annihilation();
+
+    //< Fall of Shan'bu
+    AddSC_fall_of_shan_bu();
+    AddSC_instance_fall_of_shan_bu();
+
+    //< Pursuing the Black Harvest
+    AddSC_pursing_the_black_harvest();
+    AddSC_instance_pursuing_the_black_harvest();
+
+    //< Thunder Forge
+    AddSC_thunder_forge();
+    AddSC_instance_thunder_forge();
+
+    //< Troves of the Thunder King
+    AddSC_troves_of_the_thunder_king();
+    AddSC_instance_troves_of_the_thunder_king();
+
+    //< A Brewing Storm
+    AddSC_a_brewing_storm();
+    AddSC_instance_a_brewing_storm();
+
+    //< A Little Patience
+    AddSC_a_little_patience();
+    AddSC_instance_a_little_patience();
+
+    //< Assault on Zan'vess
+    AddSC_assault_on_zanvess();
+    AddSC_instance_assault_on_zanvess();
+
+    //< Battle on the High Seas
+    AddSC_battle_on_the_high_seas();
+    AddSC_instance_battle_on_the_high_seas();
+
+    //< Blood in the Snow
+    AddSC_blood_in_the_snow();
+    AddSC_instance_blood_in_the_snow();
+
+    //< Brewmoon Festival
+    AddSC_brewmoon_festival();
+    AddSC_instance_brewmoon_festival();
+
+    //< Celestial Tournament
+    AddSC_celestial_tournament();
+    AddSC_instance_celestial_tournament();
+
+    //< Crypt of Forgotten Kings
+    AddSC_crypt_of_forgotten_kings();
+    AddSC_instance_crypt_of_forgotten_kings();
+
+    //< Dagger in the Dark
+    AddSC_dagger_in_the_dark();
+    AddSC_instance_dagger_in_the_dark();
+
+    //< Dark Heart of Pandaria
+    AddSC_dark_heart_of_pandaria();
+    AddSC_instance_dark_heart_of_pandaria();
+
+    //< Domination Point
+    AddSC_domination_point();
+    AddSC_instance_domination_point();
+
+    //< Greenstone Village
+    AddSC_greenstone_village();
+    AddSC_instance_greenstone_village();
+
+    //< Lion's Landing
+    AddSC_lions_landing();
+    AddSC_instance_lions_landing();
+
+    //< The Secrets of Ragefire
+    AddSC_the_secrets_of_ragefire();
+    AddSC_instance_the_secrets_of_ragefire();
+
+    //< Unga Ingoo
+    AddSC_unga_ingoo();
+    AddSC_instance_unga_ingoo();
+
+    //< Proving Grounds
+    AddSC_instance_proving_grounds();
+    AddSC_proving_grounds();
+
+    //< Broken Shore
+    AddSC_instance_broken_shore();
+    AddSC_broken_shore();
+
+    //< Shield's Reset
+    AddSC_instance_shields_rest();
+    AddSC_shields_rest();
+
+    //< Vengeance Art
+    AddSC_DH_vengeance_art_scenario();
+
+    // Rogue Art
+    AddSC_Dreadblades();
+
+    AddSC_Kingslayers();
+    AddSC_instance_kingslayers();
+
+    AddSC_Devourer();
+    AddSC_instance_devourer();
+
+    // Warrior scenarios
+    AddSC_wars_intro_scenario();
+
+    AddSC_Warswords();
+    AddSC_instance_warswords();
+
+    AddSC_EarthWarder();
+
+    // Priest Art
+    AddSC_Tuure();
+    AddSC_instance_tuure();
+
+    // Pdladin Art
+    AddSC_SilverHand();
+    AddSC_instance_silver_hand();
+
+    // Shaman Art
+    AddSC_Sharasdal();
+    AddSC_instance_sharasdal();
+
+    // Monk Art
+    AddSC_monk_intro_scenario();
+
+    AddSC_Sheylun();
+
+    // Art Common
+    AddSC_instance_class_art_scenarios();
+
+    AddSC_AssaultBrokenShore();
+    AddSC_instance_AssaultBrokenShore();
+
+    AddSC_InvasionAzsuna();
+    AddSC_instance_InvasionAzsuna();
+
+    AddSC_InvasionHM();
+    AddSC_instance_invasion_HM();
+
+    AddSC_InvasionValshara();
+    AddSC_instance_invasion_valshara();
+
+    AddSC_InvasionStormheim();
+    AddSC_instance_invasion_stormheim();
+
+    AddSC_instance_walling_caverns_pet();
+    AddSC_instance_dead_mines_per();
+
+    AddSC_army_training();
+    AddSC_instance_army_training();
+
+    AddSC_instance_temple_of_the_jade_serpent();
+
+    AddSC_brawlers_guild();
+    AddSC_brawlers_guild_bosses_rank_one();
+    AddSC_brawlers_guild_bosses_rank_two();
+    AddSC_brawlers_guild_bosses_rank_three();
+    AddSC_brawlers_guild_bosses_rank_four();
+    AddSC_brawlers_guild_bosses_rank_five();
+    AddSC_brawlers_guild_bosses_rank_six();
+    AddSC_brawlers_guild_bosses_rank_seven();
+}
